open-computer-use/requirements.txt
openai==0.27.8
anthropic==0.3.0
pyautogui==0.9.53
pillow==9.5.0
python-dotenv==1.0.0
gradio==3.30.0
e2b-desktop==1.2.0
ffmpeg-python==0.2.0
